+++ CHANGELOG.md
@@ -8,6 +8,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.4.9, 4/24/2026
+
+### Added
+
+N/A
+
+### Removed
+
+N/A
+
+### Changed
+
+1. For performance improvement, use new virtual threads to handle HTTP response stream in AsyncHttpClient.
+2. OSS update - vertx version 5.0.11, spring reactor-bom version 2025.0.5
+
+---
 ## Version 4.4.8, 4/16/2026
 
 ### Added